Rename content` folder to `HelperContent`
