+++ src/components/FloatingHelper/FloatingHelper.protractor.driver.ts
@@ -1,7 +1,7 @@
 import {$, ElementFinder, by} from 'protractor';
 import {popoverDriverFactory, PopoverDriver} from 'wix-ui-core/dist/src/baseComponents/Popover/Popover.protractor.driver';
 import {DataHooks} from './DataHooks';
-import {helperContentDriverFactory, HelperContentDriver} from './content/HelperContent.protractor.driver';
+import {helperContentDriverFactory, HelperContentDriver} from '../../components/FloatingHelper/HelperContent/HelperContent.protractor.driver';
 import {DriverFactory} from 'wix-ui-core/dist/src/common/BaseDriver.protractor';
 
 export interface FloatingHelperDriver extends PopoverDriver {
