+++ src/components/FloatingHelper/HelperContent/HelperContent.protractor.driver.ts
@@ -1,7 +1,7 @@
 import {ElementFinder} from 'protractor';
 import {DataHooks} from './DataHooks';
 import {DriverFactory, BaseDriver} from 'wix-ui-core/dist/src/common/BaseDriver.protractor';
-import {textDriverFactory} from '../../Text/Text.protractor.driver';
+import {textDriverFactory} from '../../../components/Text/Text.protractor.driver';
 
 export interface HelperContentDriver extends BaseDriver {
   hasTitle: () => Promise<boolean>;
