+++ src/components/FloatingHelper/HelperContent/HelperContent.spec.tsx
