+++ src/components/FloatingHelper/HelperContent/HelperContent.st.css
